Fixing bugs, added track caching.

+++ app/src/main/c/Track.h
@@ -1,7 +1,7 @@
 
 struct Track
 {
-	int Id;
+	long long Id;
 	char Path[512];
 	char Title[256];
 	char Author[256];
